+++ HISTORY.md
@@ -1,5 +1,8 @@
 History of features added to LiChess Tools in time (reversed order)
 
+16 Feb
+  - sequential and spaced repetition flow for Interactive Lessons
+
 15 Feb
   - Search by number of plys, reduce number of plys